Changes in 1.2.1:
- adjusted login details

- there was problem with copying the url to the windows clipboard: Workaround is to use Clipboard.SetDataObject instead of Clipboard.SetText

As the clipboard is shared by all UI applications, you will run into this from time to time, obviously you don't want your application to crash if it failed to write to the clipboard, so gracefully handling ExternalException is reasonable, i would suggest presenting and error to the user if the setobjectdata call to write to the clipboard fails.

A suggestion would be to use (via p/invoke) user32!GetOpenClipboardWindow to see if another application has the clipboard open, it will return the HWND of the window which has the clipboard open, or IntPtr.Zero if no application had it open, you could spin on the value until it's IntPtr.Zero for a specified amount of time.